Move EnumValueParser to PropertyValueParser because all value parsers live there
